+++ plugin-server/src/main/ingestion-queues/on-event-handler-consumer.ts
@@ -1,20 +1,10 @@
-import { Consumer, Kafka } from 'kafkajs'
+import { Consumer } from 'kafkajs'
 
 import { KAFKA_EVENTS_JSON, prefix as KAFKA_PREFIX } from '../../config/kafka-topics'
-import { Hub, PluginServerService, PluginsServerConfig } from '../../types'
-import { PostgresRouter } from '../../utils/db/postgres'
+import { Hub, PluginServerService } from '../../types'
 import { status } from '../../utils/status'
-import { ActionManager } from '../../worker/ingestion/action-manager'
-import { ActionMatcher } from '../../worker/ingestion/action-matcher'
-import { AppMetrics } from '../../worker/ingestion/app-metrics'
-import { GroupTypeManager } from '../../worker/ingestion/group-type-manager'
-import { HookCommander } from '../../worker/ingestion/hooks'
-import { OrganizationManager } from '../../worker/ingestion/organization-manager'
-import { TeamManager } from '../../worker/ingestion/team-manager'
-import { RustyHook } from '../../worker/rusty-hook'
 import { eachBatchAppsOnEventHandlers } from './batch-processing/each-batch-onevent'
-import { eachBatchWebhooksHandlers } from './batch-processing/each-batch-webhooks'
-import { KafkaJSIngestionConsumer, setupEventHandlers } from './kafka-queue'
+import { KafkaJSIngestionConsumer } from './kafka-queue'
 
 export const startAsyncOnEventHandlerConsumer = async ({
     hub, // TODO: remove needing to pass in the whole hub and be more selective on dependency injection.
@@ -32,100 +22,12 @@ export const startAsyncOnEventHandlerConsumer = async ({
 
     const queue = buildOnEventIngestionConsumer({ hub })
 
-    await hub.actionManager.start()
     await queue.start()
 
     return {
         id: 'on-event-ingestion',
         healthcheck: makeHealthCheck(queue.consumer, queue.sessionTimeout),
         onShutdown: async () => await queue.stop(),
-    }
-}
-
-export const startAsyncWebhooksHandlerConsumer = async ({
-    kafka, // TODO: remove needing to pass in the whole hub and be more selective on dependency injection.
-    postgres,
-    teamManager,
-    organizationManager,
-    actionMatcher,
-    actionManager,
-    serverConfig,
-    rustyHook,
-    appMetrics,
-    groupTypeManager,
-}: {
-    kafka: Kafka
-    postgres: PostgresRouter
-    teamManager: TeamManager
-    organizationManager: OrganizationManager
-    serverConfig: PluginsServerConfig
-    rustyHook: RustyHook
-    appMetrics: AppMetrics
-    groupTypeManager: GroupTypeManager
-    actionMatcher: ActionMatcher
-    actionManager: ActionManager
-}): Promise<PluginServerService> => {
-    /*
-        Consumes analytics events from the Kafka topic `clickhouse_events_json`
-        and processes any onEvent plugin handlers configured for the team.
-
-        At the moment this is just a wrapper around `IngestionConsumer`. We may
-        want to further remove that abstraction in the future.
-    */
-    status.info('🔁', `Starting webhooks handler consumer`)
-
-    const consumer = kafka.consumer({
-        // NOTE: This should never clash with the group ID specified for the kafka engine posthog/ee/clickhouse/sql/clickhouse.py
-        groupId: `${KAFKA_PREFIX}clickhouse-plugin-server-async-webhooks`,
-        sessionTimeout: serverConfig.KAFKA_CONSUMPTION_SESSION_TIMEOUT_MS,
-        rebalanceTimeout: serverConfig.KAFKA_CONSUMPTION_REBALANCE_TIMEOUT_MS ?? undefined,
-        readUncommitted: false,
-    })
-    setupEventHandlers(consumer)
-
-    const hookCannon = new HookCommander(
-        postgres,
-        teamManager,
-        organizationManager,
-        rustyHook,
-        appMetrics,
-        serverConfig.EXTERNAL_REQUEST_TIMEOUT_MS
-    )
-    const concurrency = serverConfig.TASKS_PER_WORKER || 20
-
-    await actionManager.start()
-    await consumer.subscribe({ topic: KAFKA_EVENTS_JSON, fromBeginning: false })
-    await consumer.run({
-        eachBatch: (payload) =>
-            eachBatchWebhooksHandlers(
-                payload,
-                actionMatcher,
-                hookCannon,
-                concurrency,
-                groupTypeManager,
-                organizationManager,
-                postgres
-            ),
-    })
-
-    const onShutdown = async () => {
-        await actionManager.stop()
-        try {
-            await consumer.stop()
-        } catch (e) {
-            status.error('🚨', 'Error stopping consumer', e)
-        }
-        try {
-            await consumer.disconnect()
-        } catch (e) {
-            status.error('🚨', 'Error disconnecting consumer', e)
-        }
-    }
-
-    return {
-        id: 'webhooks-ingestion',
-        healthcheck: makeHealthCheck(consumer, serverConfig.KAFKA_CONSUMPTION_SESSION_TIMEOUT_MS),
-        onShutdown,
     }
 }
