+++ plugin-server/src/worker/ingestion/team-manager.ts
@@ -164,7 +164,6 @@ export async function fetchTeam(client: PostgresRouter, teamId: Team['id']): Pro
                 name,
                 anonymize_ips,
                 api_token,
-                slack_incoming_webhook,
                 session_recording_opt_in,
                 person_processing_opt_out,
                 heatmaps_opt_in,
@@ -200,7 +199,6 @@ export async function fetchTeamByToken(client: PostgresRouter, token: string): P
                 name,
                 anonymize_ips,
                 api_token,
-                slack_incoming_webhook,
                 session_recording_opt_in,
                 person_processing_opt_out,
                 heatmaps_opt_in,
